PTFM-15867 ua --test flag
Summary:
Added an optional flag to the upload agent to provide the user with a quick test to make sure
their settings are correct. The test includes checking the current versions used, operating system info,
proxy settings, a call to whoami to test connectivity to the DNAnexus api, the location of the certificate file
used, resolving amazon s3 and connecting to google.com.

Test Plan: Current upload agent tests.

Subscribers: bagee

Differential Revision: https://dnanexus.phacility.com/D1158